applied solaris8/netbsd/other fixes patch by Jürgen Keil <jk@tools.de>
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@1039 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ mixer.h
@@ -7,11 +7,11 @@
 extern int    mixer_usemaster;
 extern char * mixer_device;
 
-extern void mixer_getvolume( int *l,int *r );
-extern void mixer_setvolume( int l,int r );
+extern void mixer_getvolume( float *l,float *r );
+extern void mixer_setvolume( float l,float r );
 extern void mixer_incvolume( void );
 extern void mixer_decvolume( void );
-extern int  mixer_getbothvolume( void );
+extern float mixer_getbothvolume( void );
 
 //extern void mixer_setbothvolume( int v );
 #define mixer_setbothvolume( v ) mixer_setvolume( v,v )
